requirements.txt
corenlp-python == 3.4.1.post1
gensim == 0.13.4.1
numpy  == 1.11.3
pandas == 0.19.2
progressbar2  == 3.12.0
tensorflow == 0.12.1
stanford-corenlp-pywrapper == 0.1.0
pickleshare == 0.7.4
scipy == 0.18.1